# contracts/openapi/agent-v1.openapi.json
{
  "info": {
    "title": "AI-native PaaS Agent API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mcp/v1/invoke": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../mcp/v1/invocation-request.schema.json"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "synchronous result"
          },
          "202": {
            "description": "asynchronous operation"
          },
          "4XX": {
            "description": "stable public error"
          }
        }
      }
    },
...
  }
}